requirements.txt
alembic==1.13.2
appnope==0.1.4
asttokens==2.4.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.1
openai==0.6.0
packaging==24.1
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.17.0
python-pptx==1.0.0
pyzmq==26.0.3
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
XlsxWriter==3.2.0
